Fix compiler warnings regarding tp_print type.
git-svn-id: svn://svn.code.sf.net/p/crossfire/code/server/trunk@21833 282e977c-c81d-0410-88c4-b93c2d0d6712

+++ plugins/cfpython/cfpython_map.c
@@ -560,7 +560,7 @@ PyTypeObject Crossfire_MapType = {
     sizeof(Crossfire_Map),     /* tp_basicsize*/
     0,                         /* tp_itemsize*/
     Crossfire_Map_dealloc,     /* tp_dealloc*/
-    NULL,                      /* tp_print*/
+    (long int)NULL,            /* tp_print*/
     NULL,                      /* tp_getattr*/
     NULL,                      /* tp_setattr*/
 #ifdef IS_PY3K
